// packages/clapio/src/fixtures/pets.gen.ts
export const spec = {
  document: {
    "openapi": "3.1.0",
    "info": {
      "title": "Pets",
      "version": "1.0.0"
    },
    "paths": {
      "/stores/{storeId}/pets": {
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
// ...
      },
// ...
    },
// ...
  },
// ...
} satisfies Clapio.Spec;